kernel builder: generate kernel args in builder.construct_kernel
